+++ R/boilerplate-nnetar.R
@@ -57,7 +57,8 @@
 #'   .value_col = value,
 #'   .rsamp_obj = splits,
 #'   .formula = value ~ .,
-#'   .grid_size = 5
+#'   .grid_size = 5,
+#'   .tune = FALSE
 #' )
 #'
 #' ts_nnetar$recipe_info