+++ R/boilerplate-prophet-xgboost.R
@@ -58,7 +58,8 @@
 #'   .value_col = value,
 #'   .rsamp_obj = splits,
 #'   .formula = value ~ .,
-#'   .grid_size = 5
+#'   .grid_size = 5,
+#'   .tune = FALSE
 #' )
 #'
 #' ts_prophet_boost$recipe_info