+++ R/boilerplate-svm-rbf.R
@@ -59,7 +59,8 @@
 #'   .value_col = value,
 #'   .rsamp_obj = splits,
 #'   .formula = value ~ .,
-#'   .grid_size = 3
+#'   .grid_size = 3,
+#'   .tune = FALSE
 #' )
 #'
 #' ts_auto_rbf$recipe_info