+++ NEWS.md
@@ -7,6 +7,7 @@ None
 1. Fix #459 - Add function `ts_growth_rate_vec()`
 2. Fix #463 - Add function `ts_adf_test()`
 3. Fix #417 - Add function `auto_stationarize()`
+4. Fix #460 - Add function `ts_growth_rate_augment()`
 
 ## Minor Fixes and Improvements
 1. Fix #456 Fix boilerplate examples to set the `.true` param to `FALSE`